+++ src/components/modules/tools.ts
@@ -26,6 +26,8 @@ import MoveUpTune from "../block-tunes/block-tune-move-up";
  * Creates Instances from Plugins and binds external config to the instances
  */
 
+type ToolClass = BlockTool | InlineTool | BlockTune;
+
 /**
  * Class properties:
  *
@@ -50,7 +52,7 @@ export default class Tools extends Module {
    *
    * @returns {object<Tool>}
    */
-  public get available(): Map<string, InlineTool | BlockTool | BlockTune> {
+  public get available(): Map<string, ToolClass> {
     return this.toolsAvailable;
   }
 
@@ -59,7 +61,7 @@ export default class Tools extends Module {
    *
    * @returns {Tool[]}
    */
-  public get unavailable(): Map<string, InlineTool | BlockTool | BlockTune> {
+  public get unavailable(): Map<string, ToolClass> {
     return this.toolsUnavailable;
   }
 
@@ -150,12 +152,12 @@ export default class Tools extends Module {
   /**
    * Tools` classes available to use
    */
-  private readonly toolsAvailable: Map<string, InlineTool | BlockTool | BlockTune> = new Map();
+  private readonly toolsAvailable: Map<string, ToolClass> = new Map();
 
   /**
    * Tools` classes not available to use because of preparation failure
    */
-  private readonly toolsUnavailable: Map<string, InlineTool | BlockTool | BlockTune> = new Map();
+  private readonly toolsUnavailable: Map<string, ToolClass> = new Map();
 
   /**
    * Cache for the prepared inline tools
@@ -164,23 +166,18 @@ export default class Tools extends Module {
    * @private
    */
   private _inlineTools: Map<string, InlineTool> = null;
-  private _blockTools: Map<string, BlockTool> = null;
 
   /**
-   * @class
-   *
-   * @param {EditorConfig} config - Editor's configuration
+   * Cache for the prepared block tools
    */
-  constructor({ config }) {
-    super({ config });
-  }
+  private _blockTools: Map<string, BlockTool> = null;
 
   /**
    * Returns internal tools
    *
    * @param type - if passed, Tools will be filtered by type
    */
-  public getInternal(type?: ToolType): Map<string, BlockTool | InlineTool | BlockTune> {
+  public getInternal(type?: ToolType): Map<string, ToolClass> {
     let tools = Array
       .from(this.available.entries())
       .filter(([, tool]) => {