+++ src/components/tools/base.ts
@@ -11,27 +11,37 @@ export enum UserSettings {
   Config = 'config',
 }
 
-export enum InternalSettings {
-  IsEnabledLineBreaks = 'enableLineBreaks',
-  IsInline = 'isInline',
-  IsTune = 'isTune',
-  Title = 'title', // for Inline Tools. Block Tools can pass title along with icon through the 'toolbox' static prop.
+export enum CommonInternalSettings {
   Shortcut = 'shortcut',
-  Toolbox = 'toolbox',
   SanitizeConfig = 'sanitize',
+
+}
+
+export enum InternalBlockToolSettings {
+  IsEnabledLineBreaks = 'enableLineBreaks',
+  Toolbox = 'toolbox',
   ConversionConfig = 'conversionConfig',
   IsReadOnlySupported = 'isReadOnlySupported',
   PasteConfig = 'pasteConfig'
 }
 
-export type ToolConfig = Omit<ToolSettings, 'class'>
+export enum InternalInlineToolSettings {
+  IsInline = 'isInline',
+  Title = 'title', // for Inline Tools. Block Tools can pass title along with icon through the 'toolbox' static prop.
+}
+
+export enum InternalTuneSettings {
+  IsTune = 'isTune',
+}
+
+export type ToolOptions = Omit<ToolSettings, 'class'>
 
 interface ConstructorOptions {
   name: string;
   constructable: ToolConstructable;
-  config: ToolConfig;
+  config: ToolOptions;
   api: API;
-  defaultTool: string;
+  isDefault: boolean;
   isInternal: boolean;
   defaultPlaceholder?: string | false;
 }
@@ -51,9 +61,14 @@ export default abstract class BaseTool<Type extends Tool> {
   public name: string;
 
   /**
-   * Flag show is current Tool internal or not
+   * Flag show is current Tool internal (bundled with EditorJS core) or not
    */
   public readonly isInternal: boolean;
+
+  /**
+   * Flag show is current Tool default or not
+   */
+  public readonly isDefault: boolean;
 
   /**
    * EditorJS API for current Tool
@@ -63,17 +78,12 @@ export default abstract class BaseTool<Type extends Tool> {
   /**
    * Current tool user configuration
    */
-  protected config: ToolConfig;
+  protected config: ToolOptions;
 
   /**
    * Tool's constructable blueprint
    */
   protected constructable: ToolConstructable;
-
-  /**
-   * Editor default tool
-   */
-  protected defaultTool: string;
 
   /**
    * Default placeholder specified in EditorJS user configuration
@@ -96,7 +106,7 @@ export default abstract class BaseTool<Type extends Tool> {
     constructable,
     config,
     api,
-    defaultTool,
+    isDefault,
     isInternal = false,
     defaultPlaceholder,
   }: ConstructorOptions) {
@@ -104,7 +114,7 @@ export default abstract class BaseTool<Type extends Tool> {
     this.name = name;
     this.constructable = constructable;
     this.config = config;
-    this.defaultTool = defaultTool;
+    this.isDefault = isDefault;
     this.isInternal = isInternal;
     this.defaultPlaceholder = defaultPlaceholder;
   }
@@ -112,7 +122,7 @@ export default abstract class BaseTool<Type extends Tool> {
   /**
    * Returns Tool user configuration
    */
-  public get settings(): ToolConfig {
+  public get settings(): ToolOptions {
     const config = this.config[UserSettings.Config] || {};
 
     if (this.isDefault && !('placeholder' in config) && this.defaultPlaceholder) {
@@ -147,7 +157,7 @@ export default abstract class BaseTool<Type extends Tool> {
    * Returns shortcut for Tool (internal or specified by user)
    */
   public get shortcut(): string | undefined {
-    const toolShortcut = this.constructable[InternalSettings.Shortcut];
+    const toolShortcut = this.constructable[CommonInternalSettings.Shortcut];
     const userShortcut = this.settings[UserSettings.Shortcut];
 
     return userShortcut || toolShortcut;
@@ -157,14 +167,7 @@ export default abstract class BaseTool<Type extends Tool> {
    * Returns Tool's sanitizer configuration
    */
   public get sanitizeConfig(): SanitizerConfig {
-    return this.constructable[InternalSettings.SanitizeConfig];
-  }
-
-  /**
-   * Returns true if current Tool is default
-   */
-  public get isDefault(): boolean {
-    return this.name === this.defaultTool;
+    return this.constructable[CommonInternalSettings.SanitizeConfig];
   }
 
   /**